+++ CHANGELOG.md
@@ -5,11 +5,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Dimension flag
+- Output flag
+
+### Changes
+
+- Removed short flag for width and height
+- Removed space character in all charsets
+
+---
+
 ## [0.2.1] - 2026-03-02
 
 ### Changes
 
 - Enhance auto width and height detection and fallback
+
+---
 
 ## [0.2.0] - 2026-03-02
 
@@ -18,6 +34,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Width and height flag
 - Terminal size auto detect
 - Usage and build from source instruction
+
+---
 
 ## [0.1.0] - 2026-02-25